Refactor to use es6 modules
