+++ src/create-grid.jsx
@@ -1,11 +1,12 @@
-'use strict';
+export default function createGrid(rows, cols) {
 
-module.exports = (rows, cols) => {
   return { getCoords: getCoords }
+
   function getCoords(n) {
     return {
         x: n % cols
       , y: Math.floor(n / cols)
     }
   }
+
 }
