+++ src/index.jsx
@@ -1,0 +1,1 @@
+import './stage.jsx'
