+++ src/index.less
@@ -1,6 +1,6 @@
 @width: 64px;
 @height: @width;
-@columns: 10;
+@columns: 11;
 @rows: 10;
 
 body, html {
